requirements.txt
tensorflow~=2.4.1
PyYAML~=5.3.1
dacite~=1.6.0
setuptools~=51.0.0
matplotlib~=3.4.0